Implementation of FiniteStateMachineFactory and Automata
