+++ src/main/java/org/aksw/simba/squirrel/analyzer/mime/TypeDetector.java
@@ -9,11 +9,11 @@ import java.io.File;
  * This interface defines the functionality to detect the mime-types, especially
  * of RDF serializations
  *
- * @author Abhishek Hassan Chandrashekar (abhihc@mail.uni-paderborn.de)
+ * @author Abhishek Hassan Chandrashekar&ouml;der (abhihc@mail.uni-paderborn.de)
  *
  */
 
 public interface TypeDetector {
 
-    public Lang detectMimeType(File data);
+	public Lang detectMimeType(File data);
 }
